refactor: 기타 css 및 클래스명 수정
