+++ src/pages/Login.js
@@ -2,7 +2,7 @@ import "../styles/member.scss";
 import { useForm } from "react-hook-form";
 import { ErrorMessage } from "@hookform/error-message";
 import { Link } from "react-router-dom";
-import { AuthLogin } from "../apis/AuthApi";
+import { AuthLogin } from "../utils/AuthApi";
 import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import { useCookies } from "react-cookie";
@@ -43,7 +43,7 @@ const Login = () => {
   return (
     <div className="form-wrapper">
       <div className="form-wrapperInner">
-        <h1 className="title">로그인</h1>
+        <h1 className="form__title">로그인</h1>
         <form onSubmit={handleSubmit(onSubmit)} className="form">
           <div className="inputField">
             <input
