+++ src/styles/layout.scss
@@ -3,5 +3,5 @@
 .content {
   max-width: util.rem(1120px);
   margin: 0 auto;
-  padding: util.rem(48px) 0;
+  padding: util.rem(48px) util.rem(20px);
 }
